pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "kiu"
version = "0.5.0"
description = "Knowledge in Use skill bundle schemas, validator, and candidate pipeline."
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
]
